update safety model and map controllers
