+++ server/app.js
@@ -1,22 +1,33 @@
+require('dotenv').config();
 const express = require('express');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
 
-const app = express(); 
+const mapRoutes = require('./routes/map-routes');
 
-require('dotenv').config();
+const app = express();
+app.use(bodyParser.json());
+
+app.use((req, res, next) => {
+	res.setHeader('Access-Control-Allow-Origin', '*');
+	res.setHeader(
+		'Access-Control-Allow-Headers',
+		'Origin, X-Requested-With, Content-Type, Accept, Authoriztion'
+	);
+	res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PATCH, DELETE');
+
+	next();
+});
+
+app.use('/api/map', mapRoutes);
+
 const mongoURI = process.env.MONGO_URI;
 
-// mongoose.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true })
-//   .then(() => console.log('MongoDB connected'))
-//   .catch(err => console.log(err));
-
-
-mongoose.connect(
-  mongoURI, { useNewUrlParser: true, useUnifiedTopology: true }
-).then(()=>{
-    console.log('MongoDB connected')
-    app.listen(8000);
-}).catch(err => {
-    console.log(err);
-})
+mongoose
+	.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true })
+	.then(() => {
+		app.listen(8000);
+	})
+	.catch((err) => {
+		console.log(err);
+	});
