+++ server/models/place.js
@@ -11,12 +11,12 @@ const placeSchema = new Schema({
 		lng: { type: Number, required: true },
 	},
 	box: {
-		lat1: { type: Number, required: true },
-		lng1: { type: Number, required: true },
-		lat2: { type: Number, required: true },
-		lng2: { type: Number, required: true },
+		lat1: { type: Number, required: false },
+		lng1: { type: Number, required: false },
+		lat2: { type: Number, required: false },
+		lng2: { type: Number, required: false },
 	},
-	creator: { type: mongoose.Types.ObjectId, required: true, ref: 'User' },
+	creator: { type: mongoose.Types.ObjectId, required: false, ref: 'User' },
 });
 
 module.exports = mongoose.model('Place', placeSchema);
